+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # [v0.0.31](https://github.com/opensafely-actions/cox-ipw/releases/tag/v0.0.31)
 
-- Save analysis ready now uses `foreign::write.dta` so that output can be read directly into Stata
+- Save analysis ready now uses `foreign::write.dta()` so that output can be read directly into Stata.
 
 # [v0.0.30](https://github.com/opensafely-actions/cox-ipw/releases/tag/v0.0.30)